Create continuity,add error message, change file address
Prior to this commit, the program was running only once. When the program couldn't find the word,it was return nothing.The file address was local so it had to be regulated when downloaded.

this commit add continuity and exit key. The file address is changed and no need to regulate. Add controller variable for check the word exists. Add a message that "object not found" to express not exist the word