Transform to MVC design pattern,add german dictionary, add map, add change key
Prior to this commit, program had a single Dictionary class, a words.txt. It was running on the file for every request. That's why it was slow. The program returned just meaning

this commit added four classes is that Data, WordView, WordModel, WordController and it changed Dictionary class name to Main.the program was transformed to MVC design pattern. words is extracted to vocabulary maps from files when start the program. it made faster to running "change" key was added for change dictionary. processes was separated with methods.It returns word and meaning together